Move logging_naive.h -> debug.h
