+++ src/android/main.c
@@ -1,6 +1,6 @@
 
+#include "../debug.h"
 #include "../filesys.h"
-#include "../logging_native.h"
 #include "../settings.h"
 
 #include <pthread.h>
