+++ src/av/video.c
@@ -3,7 +3,7 @@
 #include "utox_av.h"
 
 #include "../friend.h"
-#include "../logging_native.h"
+#include "../debug.h"
 #include "../macros.h"
 #include "../main_native.h"
 #include "../self.h"
