+++ src/chatlog.c
@@ -1,7 +1,7 @@
 #include "chatlog.h"
 
 #include "filesys.h"
-#include "logging_native.h"
+#include "debug.h"
 #include "main_native.h"
 #include "messages.h"
 #include "text.h"
