+++ src/cocoa/main.m
@@ -2,7 +2,7 @@
 
 #include "../commands.h"
 #include "../filesys.h"
-#include "../logging_native.h"
+#include "../debug.h"
 #include "../main.h"
 #include "../main_native.h"
 #include "../theme.h"
