+++ src/cocoa/video.m
@@ -1,7 +1,7 @@
 #include "main.h"
 
 #include "../friend.h"
-#include "../logging_native.h"
+#include "../debug.h"
 #include "../main.h"
 #include "../main_native.h"
 #include "../settings.h"
