+++ src/file_transfers.c
@@ -1,7 +1,7 @@
 #include "file_transfers.h"
 
 #include "friend.h"
-#include "logging_native.h"
+#include "debug.h"
 #include "macros.h"
 #include "main_native.h"
 #include "self.h"
