+++ src/friend.c
@@ -4,7 +4,7 @@
 #include "dns.h"
 #include "filesys.h"
 #include "flist.h"
-#include "logging_native.h"
+#include "debug.h"
 #include "main_native.h"
 #include "self.h"
 #include "settings.h"
