+++ src/groups.c
@@ -1,7 +1,7 @@
 #include "groups.h"
 
 #include "flist.h"
-#include "logging_native.h"
+#include "debug.h"
 #include "macros.h"
 #include "main_native.h"
 #include "self.h"
