+++ src/inline_video.c
@@ -13,7 +13,7 @@
 
 #include "inline_video.h"
 
-#include "logging_native.h"
+#include "debug.h"
 #include "macros.h"
 #include "main_native.h"
 #include "settings.h"
