+++ src/logging.c
@@ -1,4 +1,4 @@
-#include "logging_native.h"
+#include "debug.h"
 
 #include "settings.h"
