+++ src/main.c
@@ -4,10 +4,10 @@
 
 #include "main.h"
 
+#include "debug.h"
 #include "flist.h"
 #include "friend.h"
 #include "groups.h"
-#include "logging_native.h"
 #include "main_native.h"
 #include "settings.h"
 #include "theme.h"
