+++ src/tox.c
@@ -5,7 +5,7 @@
 #include "flist.h"
 #include "friend.h"
 #include "groups.h"
-#include "logging_native.h"
+#include "debug.h"
 #include "macros.h"
 #include "main_native.h"
 #include "self.h"
