+++ src/tox_callbacks.c
@@ -3,7 +3,7 @@
 #include "file_transfers.h"
 #include "friend.h"
 #include "groups.h"
-#include "logging_native.h"
+#include "debug.h"
 #include "macros.h"
 #include "settings.h"
 #include "text.h"
