+++ src/ui/dropdowns.c
@@ -6,7 +6,7 @@
 #include "../flist.h"
 #include "../friend.h"
 #include "../groups.h"
-#include "../logging_native.h"
+#include "../debug.h"
 #include "../macros.h"
 #include "../main_native.h"
 #include "../settings.h"
