+++ src/ui/svg.c
@@ -2,7 +2,7 @@
 
 #include "draw.h"
 
-#include "../logging_native.h"
+#include "../debug.h"
 #include "../macros.h"
 #include "../main.h"
