+++ src/updater.c
@@ -1,6 +1,6 @@
 #include "updater.h"
 
-#include "logging_native.h"
+#include "debug.h"
 #include "settings.h"
 #include "branding.h"
