+++ src/windows/filesys.c
@@ -1,6 +1,6 @@
 #include "main.h"
 
-#include "../logging_native.h"
+#include "../debug.h"
 #include "../main.h"
 #include "../settings.h"
