+++ src/windows/main.7.c
@@ -12,7 +12,7 @@
 #include "../filesys.h"
 #include "../file_transfers.h"
 #include "../friend.h"
-#include "../logging_native.h"
+#include "../debug.h"
 #include "../tox.h"
 #include "../settings.h"
