+++ src/windows/main.c
@@ -5,7 +5,7 @@
 #include "../filesys.h"
 #include "../flist.h"
 #include "../friend.h"
-#include "../logging_native.h"
+#include "../debug.h"
 #include "../macros.h"
 #include "../main.h"
 #include "../main_native.h"
