+++ src/xlib/event.c
@@ -2,7 +2,7 @@
 
 #include "../flist.h"
 #include "../friend.h"
-#include "../logging_native.h"
+#include "../debug.h"
 #include "../macros.h"
 #include "../main_native.h"
 #include "../settings.h"
