+++ src/xlib/filesys.c
@@ -5,7 +5,7 @@
 #include "../chatlog.h"
 #include "../file_transfers.h"
 #include "../friend.h"
-#include "../logging_native.h"
+#include "../debug.h"
 #include "../tox.h"
 #include "../settings.h"
