+++ src/xlib/gtk.c
@@ -1,9 +1,9 @@
 #include "../chatlog.h"
+#include "../debug.h"
 #include "../file_transfers.h"
 #include "../filesys.h"
 #include "../flist.h"
 #include "../friend.h"
-#include "../logging_native.h"
 #include "../macros.h"
 #include "../main_native.h"
 #include "../text.h"
